The Country Component is functionall but missing some feautures
The country component and the entire App are now complete and fully functionall it acomplishes all the major goals . first it renders all the ountries and their most basic details. it has the Search and filter by region function . it's missing three feautures and after implementing these three feautures , the App will be fully complete. first and the most important is making the App responsive,second is the renering the boreder countries names in the chosen country profile and the ability to change the chosen country to the border country that the user chosses, and the third is adding the currencies and languages sections to the chosen country details sections